refactor: share Buzz membership snapshot queries
